Recover Dreamcast CodeView source types (#17)
* Recover Dreamcast source types and track audit exceptions

* Batch Dreamcast audit dispositions

* Batch Dreamcast type audit dispositions

* Resolve exact CodeView type differences

* Batch remaining Dreamcast ABI and spell types

* Batch Dreamcast source-type recoveries

* Recover exact adventure screen update syntax

* Recover two lost exact functions

* Normalize hero class enumerators

+++ src/bitmap16.cpp
@@ -405,7 +405,7 @@ void Bitmap16Bit::darken(int x, int y, int w, int h, Bitmap816* mask,
         h = m_height - y;
 
     if (w && h) {
-        unsigned long shiftMask =
+        unsigned int shiftMask =
             ((g_colorMaskRed >> 1) & g_colorMaskRed)
             | ((g_colorMaskGreen >> 1) & g_colorMaskGreen)
             | ((g_colorMaskBlue >> 1) & g_colorMaskBlue);
